server/requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
loguru>=0.7.0
pystray
Pillow
av
numpy
onnxruntime-directml
